requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.2.4
idna==3.3
requests==2.28.1
six==1.16.0
urllib3==1.26.9
